% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece.plot.R
\name{ece.plot}
\alias{ece.plot}
\title{Create an empirical cross-entropy (ECE) plot}
\usage{
ece.plot(
  log.lrs,
  labels,
  method = c("binned", "smoothed"),
  nbins = 10,
  bandwidth = 0.05,
  resolution = 200
)
}
\arguments{
\item{nbins}{}
}
\value{
a ggplot2 plot showing an ECE plot.
}
\description{
Create an empirical cross-entropy (ECE) plot
}
